Implement support for --chmod and --chown rsync flags.
By default, afl-sync behaves the same as before, running rsync in
archive mode (-a) and trasferring local permissions and owner/group.
However, --chmod PERMS and --chown USER:GROUP can now be specified,
which passes these flags to rsync, enabling the user to specify the
permissions, owner and group of the remote files.

This is useful when ensuring the remote storage dir is readable and/or
writable on the server for all processes that need to access it.

The new flags are only supported when pushing and in the push phase of
syncing, not pulling. The support for the flags was implemented by
the introduction of a new parameter for AflRsync, which represents the
options for rsync, both its get and put modes.