Prevent local globbing when --chown is passed to rsync.
rsync seems to internally convert --chown=foo:bar to --usermap=*:foo
--groupmap=*:bar. Due to a bug[*], it passes this to a shell at some
point without properly escaping it. The shell then globs the wildcard,
which is not correct and likely makes the shell complain about
nonexistent matches. A workaround is to pass --protect-args to rsync
which makes makes it work as intended but may have other consequences.

[*] https://bugzilla.samba.org/show_bug.cgi?id=10705